Re-org keys and ecdsa mods - pt.3
